refactor: fold deposit and redeem components into pages
