+++ src/lib/wallet/index.ts
@@ -4,11 +4,9 @@ export { default as TokenBalance } from './TokenBalance.svelte';
 export { default as VaultBalance } from './VaultBalance.svelte';
 export { default as WalletAddress } from './WalletAddress.svelte';
 export { default as WalletBalance } from './WalletBalance.svelte';
-export { default as WalletDeposit } from './WalletDeposit.svelte';
 export { default as WalletInfo } from './WalletInfo.svelte';
 export { default as WalletInfoItem } from './WalletInfoItem.svelte';
 export { default as WalletSummary } from './WalletSummary.svelte';
 export { default as WalletTile } from './WalletTile.svelte';
-export { default as WalletTokenRedemption } from './WalletTokenRedemption.svelte';
 export * from './client';
 export * from './utils';
