+++ src/routes/wizard/deposit/payment/+page.svelte
@@ -1,13 +1,311 @@
 <script lang="ts">
 	import wizard from '../store';
-	import { Button, WizardActions } from '$lib/components';
-	import { WalletDeposit } from '$lib/wallet';
+	import { tweened } from 'svelte/motion';
+	import { cubicOut } from 'svelte/easing';
+	import fsm from 'svelte-fsm';
+	import type { FetchBalanceResult } from '@wagmi/core';
+	import { getPublicClient, prepareWriteContract, writeContract, waitForTransaction } from '@wagmi/core';
+	import { parseUnits } from 'viem';
+	import { wallet, getExplorerUrl, WalletInfo, WalletInfoItem } from '$lib/wallet';
+	import { type SignedArguments, fetchTokenInfo, getSignedArguments } from '$lib/eth-defi/eip-3009';
+	import paymentForwarderABI from '$lib/eth-defi/abi/VaultUSDCPaymentForwarder.json';
+	import {
+		Button,
+		AlertItem,
+		AlertList,
+		CryptoAddressWidget,
+		EntitySymbol,
+		MoneyInput,
+		WizardActions
+	} from '$lib/components';
+
+	const contracts: Contracts = $wizard.data.contracts;
+	const denominationToken: FetchBalanceResult = $wizard.data.denominationToken;
+	const nativeCurrency: FetchBalanceResult = $wizard.data.nativeCurrency;
+
+	let paymentValue: MaybeNumber;
+	let errorMessage: MaybeString;
+	let transactionId: Maybe<Address>;
+
+	const progressBar = tweened(0, { easing: cubicOut });
+
+	async function authorizeTransfer() {
+		const tokenInfo = await fetchTokenInfo(denominationToken.address);
+		const value = parseUnits(`${paymentValue}`, tokenInfo.decimals);
+
+		return getSignedArguments(
+			'TransferWithAuthorization',
+			$wizard.data.chainId,
+			tokenInfo,
+			$wallet.address!,
+			contracts.payment_forwarder,
+			value
+		);
+	}
+
+	async function confirmPayment(signedArgs: SignedArguments) {
+		const { request } = await prepareWriteContract({
+			address: contracts.payment_forwarder,
+			abi: paymentForwarderABI,
+			functionName: 'buySharesOnBehalfUsingTransferWithAuthorization',
+			args: [...signedArgs, 1]
+		});
+
+		return writeContract(request);
+	}
+
+	const payment = fsm('initial', {
+		initial: {
+			// restore state on wizard back/next navigation
+			restore(state) {
+				({ errorMessage, transactionId, paymentValue } = $wizard.data);
+				if (state === 'authorizing' || state === 'confirming') {
+					errorMessage = `Wallet request state lost due to window navigation;
+						please cancel wallet request and try again.`;
+					return 'failed';
+				}
+				return state;
+			},
+
+			authorize() {
+				authorizeTransfer().then(payment.confirm).catch(payment.fail);
+				return 'authorizing';
+			}
+		},
+
+		authorizing: {
+			confirm(signedArgs) {
+				confirmPayment(signedArgs).then(payment.process).catch(payment.fail);
+				return 'confirming';
+			},
+
+			fail(err) {
+				console.error('authorizeTransfer error:', err);
+				errorMessage = `Authorization to transfer ${denominationToken.symbol} tokens from your wallet account failed.`;
+				return 'failed';
+			}
+		},
+
+		confirming: {
+			process({ hash }) {
+				transactionId = hash;
+				wizard.updateData({ transactionId });
+				return 'processing';
+			},
+
+			fail(err) {
+				console.error('confirmPayment error:', err);
+				errorMessage = 'Payment transaction confirmation from wallet account failed.';
+				return 'failed';
+			}
+		},
+
+		processing: {
+			_enter({ event }) {
+				const hash = transactionId!;
+				let duration = 20_000;
+
+				if (event === 'restore') {
+					// try fetching receipt in case transaction already completed
+					getPublicClient()
+						.getTransactionReceipt({ hash })
+						.then(payment.finish)
+						.catch(() => {});
+					progressBar.set(50, { duration: 0 });
+					duration *= 0.5;
+				}
+
+				// wait for pending transaction
+				waitForTransaction({ hash }).then(payment.finish).catch(payment.fail);
+				progressBar.set(100, { duration });
+			},
+
+			_exit() {
+				progressBar.set(100, { duration: 100 });
+			},
+
+			finish(receipt) {
+				if (receipt.status !== 'success') {
+					console.error('waitForTransaction reverted:', receipt);
+					errorMessage = 'Transaction execution reverted. See blockchain explorer for details.';
+					return 'failed';
+				}
+				return 'completed';
+			},
+
+			fail(err) {
+				console.error('waitForTransaction error:', err);
+				if (err.name === 'CallExecutionError') {
+					errorMessage = `${err.shortMessage} See blockchain explorer for details.`;
+				} else {
+					errorMessage = 'Unable to verify transaction status. See blockchain explorer for details.';
+				}
+				return 'failed';
+			}
+		},
+
+		failed: {
+			_enter() {
+				wizard.updateData({ errorMessage });
+			},
+
+			retry() {
+				return transactionId ? 'processing' : 'initial';
+			}
+		},
+
+		completed: {
+			_enter({ event }) {
+				if (event === 'restore') {
+					progressBar.set(100, { duration: 0 });
+				}
+				wizard.complete('payment');
+			}
+		}
+	});
+
+	payment.restore($wizard.data.paymentState);
+	$: wizard.updateData({ paymentState: $payment });
 </script>
 
-<WalletDeposit {wizard} />
+<div class="wallet-deposit">
+	<section>
+		<h3>Your current balance</h3>
+
+		<WalletInfo alignValues="right">
+			<WalletInfoItem>
+				<EntitySymbol
+					slot="label"
+					type="token"
+					label={nativeCurrency.symbol}
+					slug={nativeCurrency.symbol.toLowerCase()}
+				/>
+				{nativeCurrency.formatted}
+			</WalletInfoItem>
+
+			<WalletInfoItem>
+				<EntitySymbol
+					slot="label"
+					type="token"
+					label={denominationToken.symbol}
+					slug={denominationToken.symbol.toLowerCase()}
+				/>
+				{denominationToken.formatted}
+			</WalletInfoItem>
+		</WalletInfo>
+	</section>
+
+	<section>
+		<h3>Enter amount to deposit</h3>
+		<form class="payment-form" on:submit|preventDefault={payment.authorize}>
+			<MoneyInput
+				bind:value={paymentValue}
+				size="xl"
+				tokenUnit={denominationToken.symbol}
+				disabled={$payment !== 'initial'}
+				on:change={() => wizard.updateData({ paymentValue })}
+			/>
+
+			{#if $payment === 'initial'}
+				<Button submit disabled={!paymentValue}>Make payment</Button>
+
+				<AlertList size="sm" status="warning">
+					<AlertItem title="Notice">
+						Investing in crypto trading carries significant risks. Past performance is not indicative of future results.
+						Only invest funds you are willing to lose.
+					</AlertItem>
+				</AlertList>
+			{/if}
+
+			{#if $payment === 'authorizing'}
+				<AlertList size="sm" status="warning">
+					<AlertItem title="Authorize transfer">
+						Please authorize the transfer of {denominationToken.symbol} tokens from your wallet account.
+					</AlertItem>
+				</AlertList>
+			{/if}
+
+			{#if $payment === 'confirming'}
+				<AlertList size="sm" status="warning">
+					<AlertItem title="Confirm transaction">
+						Please confirm the transaction in your wallet in order submit your payment.
+					</AlertItem>
+				</AlertList>
+			{/if}
+
+			{#if transactionId}
+				<div class="transaction-id">
+					<h3>Transaction ID</h3>
+					<CryptoAddressWidget address={transactionId} href={getExplorerUrl($wallet.chain, transactionId)} />
+				</div>
+
+				<progress max="100" value={$progressBar} />
+			{/if}
+
+			{#if $payment === 'processing'}
+				<AlertList size="sm" status="info">
+					<AlertItem title="Payment processing">
+						The duration of processing may vary based on factors such as blockchain congestion and gas specified. Click
+						the transaction ID above to view the status in the blockchain explorer.
+					</AlertItem>
+				</AlertList>
+			{/if}
+
+			{#if $payment === 'failed'}
+				<AlertList size="sm" status="error">
+					<AlertItem title="Error">
+						{errorMessage}
+						<Button slot="cta" size="sm" label="Try again" on:click={payment.retry} />
+					</AlertItem>
+				</AlertList>
+			{/if}
+
+			{#if $payment === 'completed'}
+				<AlertList size="sm" status="success">
+					<AlertItem title="Payment completed">
+						Your transaction completed successfully and the shares have been added to your wallet. Click "Next" below to
+						review your share balance.
+					</AlertItem>
+				</AlertList>
+			{/if}
+		</form>
+	</section>
+</div>
 
 <WizardActions>
 	<Button ghost label="Cancel" href={$wizard.returnTo} disabled={$wizard.completed.has('payment')} />
 	<Button secondary label="Back" href="balance" />
 	<Button label="Next" href="success" disabled={!$wizard.completed.has('payment')} />
 </WizardActions>
+
+<style lang="postcss">
+	.wallet-deposit {
+		display: grid;
+		gap: var(--space-3xl);
+
+		& section {
+			display: grid;
+			gap: var(--space-ls);
+		}
+
+		& h3 {
+			color: hsla(var(--hsl-text-light));
+			font: var(--f-ui-lg-medium);
+		}
+
+		& .payment-form {
+			display: grid;
+			gap: var(--space-xl);
+		}
+
+		& progress {
+			width: 100%;
+		}
+
+		& .transaction-id {
+			display: grid;
+			gap: var(--space-md);
+			justify-content: start;
+		}
+	}
+</style>
